// backend/src/types/code/allocation.ts
export enum AllocationErrorCode {
  ALLOCATION_CREATE_FAILED = "ALLOCATION_CREATE_FAILED",
  ALLOCATION_FETCH_FAILED = "ALLOCATION_FETCH_FAILED",
  ALLOCATION_NOT_FOUND = "ALLOCATION_NOT_FOUND",
  VEHICLE_ALREADY_EXISTS = "VEHICLE_ALREADY_EXISTS",
  SPOTS_NOT_FOUND = "SPOTS_NOT_FOUND",
  SPOTS_FETCH_FAILED = "SPOTS_FETCH_FAILED",
}