// backend/src/types/code/parkingCode.ts
export enum ParkingErrorCode {
  PARKING_CREATE_FAILED = "PARKING_CREATE_FAILED",
  EMAIL_ALREADY_EXISTS = "EMAIL_ALREADY_EXISTS",
  PARKING_NOT_FOUND = "PARKING_NOT_FOUND",
  PARKING_FETCH_FAILED = "PARKING_FETCH_FAILED",
  PARKING_REMOVE_FAILED = "PARKING_REMOVE_FAILED",
  PARKING_INTERNAL_SERVER_ERROR = "PARKING_INTERNAL_SERVER_ERROR"
}